# Allocate Chain entries lazily