Since many applications are using bare oss sidekiq without any middlewares, this pr can help to save time and memory.

I have tested using `bin/sidekiqload`.

### Using `memory_profiler`, I got:
**Before**:
```
Total allocated: 999716276 bytes (10908573 objects)
Total retained:  12844675 bytes (1634 objects)
```

**After**:
```
Total allocated: 951803096 bytes (10109336 objects)
Total retained:  12826021 bytes (1330 objects)
```
Speed up in memory = `(999716276  - 951803096) / 999716276.0 * 100 ~ 5%`


### Using `allocation_tracer`, I got:
**Before**:
```
 ["/Users/fatkodima/Desktop/oss/sidekiq/lib/sidekiq/middleware/chain.rb",
  81,
  :T_ARRAY]=>[2, 0, 0, 0, 0, 0],
 ["/Users/fatkodima/Desktop/oss/sidekiq/lib/sidekiq/client.rb",
  216,
  :T_ARRAY]=>[300009, 0, 0, 0, 0, 0],
 ["/Users/fatkodima/Desktop/oss/sidekiq/lib/sidekiq/middleware/chain.rb",
  118,
  :T_IMEMO]=>[200007, 0, 0, 0, 0, 0],
 ["/Users/fatkodima/Desktop/oss/sidekiq/lib/sidekiq/middleware/chain.rb",
  118,
  :T_ARRAY]=>[200006, 0, 0, 0, 0, 0],
 ["/Users/fatkodima/Desktop/oss/sidekiq/lib/sidekiq/middleware/chain.rb",
  126,
  :T_ARRAY]=>[200006, 0, 0, 0, 0, 0],
 ["/Users/fatkodima/Desktop/oss/sidekiq/lib/sidekiq/middleware/chain.rb",
  127,
  :T_IMEMO]=>[400013, 0, 0, 0, 0, 0],
 ["/Users/fatkodima/Desktop/oss/sidekiq/lib/sidekiq/middleware/chain.rb",
  127,
  :T_DATA]=>[400012, 0, 0, 0, 0, 0]
.......
```

**After**:
All those lines roughly creating 1_700_000 objects were gone away.


### Time measurements:
**Before** (avg result):
```
bin/sidekiqload  19.94s user 5.34s system 98% cpu 25.581 total
```

**After**(avg result):
```
bin/sidekiqload  18.97s user 5.14s system 98% cpu 24.278 total
```

Speed up in time = `(25.581 - 24.278) / 25.581 * 100 ~ 5%`