@mperham 2019-09-14
I don’t really like the Rails initializer. Seems unnecessary.