@fatkodima 2019-09-14
Won't threads compete for `Sidekiq`s module `@client_chain` object instantiation on first access to `Sidekiq.client_middleware`? Can something bad happen if will?