@mperham 2019-09-14
The users Sidekiq.rb initializer is the thing which will touch it and Rails boots on a single thread. Also, cli.rb touches the client middleware to log its contents, not sure if that’s done before launching threads. 

> On Sep 14, 2019, at 08:32, fatkodima <notifications@github.com> wrote:
> 
> Won't threads compete for Sidekiqs module @client_chain object instantiation on first access to Sidekiq.client_middleware? Can something bad happen if will?
> 
> —
> You are receiving this because you commented.
> Reply to this email directly, view it on GitHub, or mute the thread.