@fatkodima 2019-09-14
Yes, there is a line inside cli.rb which touches server middleware explicitly
https://github.com/mperham/sidekiq/blob/705d6ca7ea2ce73b3244b740fc0d0f8f19db8c60/lib/sidekiq/cli.rb#L71

And 2 lines, which touch both middleware chains, but only in debug mode
https://github.com/mperham/sidekiq/blob/705d6ca7ea2ce73b3244b740fc0d0f8f19db8c60/lib/sidekiq/cli.rb#L77-L78

So, probably, would need to add line for client middleware, similar to first example? 
So for sidekiq server, middlewares are partly eager loaded:
- server middleware - yes
- client middleware - only when in debug mode

> The users Sidekiq.rb initializer is the thing which will touch it and Rails boots on a single thread.

But when I have initializer like this:
```ruby
  Sidekiq.configure_client do |config|
    config.redis = { :namespace => 'myapp', :size => 25, :url => 'redis://myhost:8877/0' }
  end
```

without explicitly calling `client_middleware`, it stays not initialized. And when I have rails with puma, for example, multiple threads enqueueing jobs may compete for that object creation.

What do you think?