@fatkodima 2019-09-14
Initially I was thinking not about saving a couple of objects, but rather something like atomically `@client_chain` allocation and assignment inside `#client_middleware` before threads starting to use it. But thinking twice, looks like this is not a real problem. 
So removed rails initializer.