@mperham 2019-09-14
Nicely done, you’re doing a great job with these re-factorings.